move WEVMOS contract to contracts directory
