+++ contracts/compiled_contracts/WEVMOS.json
